Installs Cypress for e2e testing (#27)
* runs npm install cypress

* beginnings of a test file

* installs firebase-admin package

* installs cypress-firebase package

* sets up config for cypress to login with test user; refactors dev scripts to exclude test user deletion on reset

* adds cypress directory to eslintrc ignore; import order

* makes scripts more resilient

* more tweaking of dev functions and scripts

* gets basic seed-data script working and configured in beforeEach

* writes basic tests for tags feature

* makes deleteData function consistently remove all user data; adds user and menu data in seedData function

* refactors verifyUserInfo utility function to fetchUserInfo and uses it where needed

* sets up e2e command

* cleanup

* updates jest.config to ignore cypress directory

+++ scripts/dev/delete-all-accounts.js
@@ -1,6 +1,7 @@
 const admin = require('firebase-admin');
 const serviceAccount = require('../../firebase-admin-dev.json');
-const { deleteUserAccount } = require('./utility');
+const { deleteAccount } = require('./utility');
+const { TEST_UID } = require('../../cypress.env.json');
 
 admin.initializeApp({
   credential: admin.credential.cert(serviceAccount),
@@ -8,17 +9,25 @@ admin.initializeApp({
 });
 
 function deleteAllUserAccounts(nextPageToken) {
+  if (!serviceAccount.project_id.includes('dev')) {
+    throw new Error('STOP! This script is targeting the wrong environment.');
+  }
+
   admin.auth().listUsers(100, nextPageToken)
-    .then(function (listUsersResult) {
-      listUsersResult.users.forEach(function (userRecord) {
-        uid = userRecord.toJSON().uid;
-        deleteUserAccount(admin, uid);
+    .then((listUsersResult) => {
+      listUsersResult.users.forEach((userRecord) => {
+        const uid = userRecord.toJSON().uid;
+        if (uid === TEST_UID) {
+          console.log(`Skipping test user ${uid}.`);
+          return;
+        }
+        deleteAccount(admin, uid);
       });
       if (listUsersResult.pageToken) {
         deleteAllUserAccounts(listUsersResult.pageToken);
       }
     })
-    .catch(function (error) {
+    .catch((error) => {
       console.log('Error listing users:', error);
     });
 }
